+++ tests/testapp/views.py
@@ -46,7 +46,5 @@ def page_detail(request, path=None):
         return redirect(page.redirect_to_url or page.redirect_to_page)
     return render(request, page.template.template_name, {
         'page': page,
-        'contents': renderer.contents_for_item(
-            page, page.ancestors().reverse()),
-        'renderer': renderer,
+        'regions': renderer.regions(page, page.ancestors().reverse()),
     })