+++ CHANGELOG.rst
@@ -7,9 +7,8 @@ Change log
 
 - Fixed oEmbed read timeouts to not crash but retry after 60 seconds
   instead.
-- Added the ``TemplatePluginRenderer.contents_for_item`` helper for
-  returning a lazy ``Contents`` instance most useful in conjunction with
-  plugin output caching.
+- Added the ``TemplatePluginRenderer.regions`` helper and the
+  ``{⅜ render_region %}`` template tag which support caching of plugins.
 
 
 `0.12`_ (2016-10-23)